Move production types to production module
