+++ photofeed/src/commonMain/kotlin/com/asturiancoder/photofeed/feed/cache/FeedStore.kt
@@ -1,0 +1,7 @@
+package com.asturiancoder.photofeed.feed.cache
+
+import com.asturiancoder.photofeed.feed.cache.model.CachedFeed
+
+interface FeedStore {
+    fun retrieve(): CachedFeed?
+}
