+++ photofeed/src/commonMain/kotlin/com/asturiancoder/photofeed/feed/cache/model/CachedFeed.kt
@@ -1,0 +1,8 @@
+package com.asturiancoder.photofeed.feed.cache.model
+
+import com.asturiancoder.photofeed.feed.feature.FeedPhoto
+
+data class CachedFeed(
+    val feed: List<FeedPhoto>,
+    val timestamp: Long,
+)
